+++ app/src/androidTest/java/org/simple/clinic/sync/MedicalHistorySyncIntegrationTest.kt
@@ -74,7 +74,6 @@ class MedicalHistorySyncIntegrationTest {
         syncCoordinator = SyncCoordinator(),
         repository = repository,
         api = syncApi,
-        userSession = userSession,
         lastPullToken = lastPullToken,
         config = config
     )
@@ -110,13 +109,13 @@ class MedicalHistorySyncIntegrationTest {
     assertThat(repository.pendingSyncRecordCount().blockingFirst()).isEqualTo(totalNumberOfRecords)
 
     // when
-    sync.push().blockingAwait()
+    sync.push()
     clearMedicalHistoryData()
-    sync.pull().blockingAwait()
+    sync.pull()
 
     // then
     val expectedPulledRecords = records.map { it.syncCompleted() }
-    val pulledRecords = repository.recordsWithSyncStatus(SyncStatus.DONE).blockingGet()
+    val pulledRecords = repository.recordsWithSyncStatus(SyncStatus.DONE)
 
     assertThat(pulledRecords).containsAtLeastElementsIn(expectedPulledRecords)
   }
@@ -133,7 +132,7 @@ class MedicalHistorySyncIntegrationTest {
     assertThat(records).containsNoDuplicates()
 
     repository.save(records).blockingAwait()
-    sync.push().blockingAwait()
+    sync.push()
     assertThat(repository.pendingSyncRecordCount().blockingFirst()).isEqualTo(0)
 
     val modifiedRecord = records[1].hadHeartAttack()
@@ -141,15 +140,15 @@ class MedicalHistorySyncIntegrationTest {
     assertThat(repository.pendingSyncRecordCount().blockingFirst()).isEqualTo(1)
 
     // when
-    sync.pull().blockingAwait()
+    sync.pull()
 
     // then
     val expectedSavedRecords = records
         .map { it.syncCompleted() }
         .filterNot { it.patientUuid == modifiedRecord.patientUuid }
 
-    val savedRecords = repository.recordsWithSyncStatus(SyncStatus.DONE).blockingGet()
-    val pendingSyncRecords = repository.recordsWithSyncStatus(SyncStatus.PENDING).blockingGet()
+    val savedRecords = repository.recordsWithSyncStatus(SyncStatus.DONE)
+    val pendingSyncRecords = repository.recordsWithSyncStatus(SyncStatus.PENDING)
 
     assertThat(savedRecords).containsAtLeastElementsIn(expectedSavedRecords)
     assertThat(pendingSyncRecords).containsExactly(modifiedRecord)