+++ app/src/androidTest/java/org/simple/clinic/sync/PatientSyncIntegrationTest.kt
@@ -68,7 +68,6 @@ class PatientSyncIntegrationTest {
         syncCoordinator = SyncCoordinator(),
         repository = repository,
         api = syncApi,
-        userSession = userSession,
         lastPullToken = lastPullToken,
         config = config
     )
@@ -107,13 +106,13 @@ class PatientSyncIntegrationTest {
     assertThat(repository.pendingSyncRecordCount().blockingFirst()).isEqualTo(totalNumberOfRecords)
 
     // when
-    patientSync.push().blockingAwait()
+    patientSync.push()
     clearPatientData()
-    patientSync.pull().blockingAwait()
+    patientSync.pull()
 
     // then
     val expectedPulledRecords = records.map { it.syncCompleted() }
-    val pulledRecords = repository.recordsWithSyncStatus(SyncStatus.DONE).blockingGet()
+    val pulledRecords = repository.recordsWithSyncStatus(SyncStatus.DONE)
 
     assertThat(pulledRecords).containsAtLeastElementsIn(expectedPulledRecords)
   }
@@ -131,7 +130,7 @@ class PatientSyncIntegrationTest {
     assertThat(records).containsNoDuplicates()
 
     repository.save(records).blockingAwait()
-    patientSync.push().blockingAwait()
+    patientSync.push()
     assertThat(repository.pendingSyncRecordCount().blockingFirst()).isEqualTo(0)
 
     val modifiedRecord = records[1].withName("Anish Acharya")
@@ -139,15 +138,15 @@ class PatientSyncIntegrationTest {
     assertThat(repository.pendingSyncRecordCount().blockingFirst()).isEqualTo(1)
 
     // when
-    patientSync.pull().blockingAwait()
+    patientSync.pull()
 
     // then
     val expectedSavedRecords = records
         .map { it.syncCompleted() }
         .filterNot { it.patientUuid == modifiedRecord.patientUuid }
 
-    val savedRecords = repository.recordsWithSyncStatus(SyncStatus.DONE).blockingGet()
-    val pendingSyncRecords = repository.recordsWithSyncStatus(SyncStatus.PENDING).blockingGet()
+    val savedRecords = repository.recordsWithSyncStatus(SyncStatus.DONE)
+    val pendingSyncRecords = repository.recordsWithSyncStatus(SyncStatus.PENDING)
 
     assertThat(savedRecords).containsAtLeastElementsIn(expectedSavedRecords)
     assertThat(pendingSyncRecords).containsExactly(modifiedRecord)