+++ app/src/main/java/org/simple/clinic/drugs/sync/PrescriptionSyncApi.kt
@@ -1,7 +1,7 @@
 package org.simple.clinic.drugs.sync
 
-import io.reactivex.Single
 import org.simple.clinic.sync.DataPushResponse
+import retrofit2.Call
 import retrofit2.http.Body
 import retrofit2.http.GET
 import retrofit2.http.Headers
@@ -13,12 +13,12 @@ interface PrescriptionSyncApi {
   @POST("v3/prescription_drugs/sync")
   fun push(
       @Body body: PrescriptionPushRequest
-  ): Single<DataPushResponse>
+  ): Call<DataPushResponse>
 
   @Headers(value = ["X-RESYNC-TOKEN: 1"])
   @GET("v3/prescription_drugs/sync")
   fun pull(
       @Query("limit") recordsToPull: Int,
       @Query("process_token") lastPullToken: String? = null
-  ): Single<PrescriptionPullResponse>
+  ): Call<PrescriptionPullResponse>
 }