+++ app/src/main/java/org/simple/clinic/facility/FacilitySyncApi.kt
@@ -1,6 +1,6 @@
 package org.simple.clinic.facility
 
-import io.reactivex.Single
+import retrofit2.Call
 import retrofit2.http.GET
 import retrofit2.http.Headers
 import retrofit2.http.Query
@@ -12,6 +12,6 @@ interface FacilitySyncApi {
   fun pull(
       @Query("limit") recordsToPull: Int,
       @Query("process_token") lastPullToken: String? = null
-  ): Single<FacilityPullResponse>
+  ): Call<FacilityPullResponse>
 
 }