+++ app/src/main/java/org/simple/clinic/medicalhistory/sync/MedicalHistorySync.kt
@@ -2,14 +2,12 @@ package org.simple.clinic.medicalhistory.sync
 
 import com.f2prateek.rx.preferences2.Preference
 import io.reactivex.Completable
-import io.reactivex.Single
 import org.simple.clinic.medicalhistory.Answer
 import org.simple.clinic.medicalhistory.MedicalHistory
 import org.simple.clinic.medicalhistory.MedicalHistoryRepository
 import org.simple.clinic.sync.ModelSync
 import org.simple.clinic.sync.SyncConfig
 import org.simple.clinic.sync.SyncCoordinator
-import org.simple.clinic.user.UserSession
 import org.simple.clinic.util.Optional
 import javax.inject.Inject
 import javax.inject.Named
@@ -18,36 +16,27 @@ class MedicalHistorySync @Inject constructor(
     private val syncCoordinator: SyncCoordinator,
     private val repository: MedicalHistoryRepository,
     private val api: MedicalHistorySyncApi,
-    private val userSession: UserSession,
     @Named("last_medicalhistory_pull_token") private val lastPullToken: Preference<Optional<String>>,
     @Named("sync_config_frequent") private val config: SyncConfig
 ) : ModelSync {
 
-  private fun canSyncData() = userSession.canSyncData().firstOrError()
-
   override val name: String = "Medical History"
 
-  override fun sync(): Completable =
-      canSyncData()
-          .flatMapCompletable { canSync ->
-            if (canSync) {
-              Completable.mergeArrayDelayError(push(), pull())
+  override val requiresSyncApprovedUser = true
 
-            } else {
-              Completable.complete()
-            }
-          }
+  override fun sync(): Completable = Completable
+      .mergeArrayDelayError(
+          Completable.fromAction { push() },
+          Completable.fromAction { pull() }
+      )
 
-  override fun push(): Completable {
-    return syncCoordinator.push(repository, pushNetworkCall = { api.push(toRequest(it)) })
+  override fun push() {
+    syncCoordinator.push(repository, pushNetworkCall = { api.push(toRequest(it)).execute().body()!! })
   }
 
-  override fun pull(): Completable {
-    return Single
-        .fromCallable { config.batchSize }
-        .flatMapCompletable { batchSize ->
-          syncCoordinator.pull(repository, lastPullToken, batchSize) { api.pull(batchSize, it) }
-        }
+  override fun pull() {
+    val batchSize = config.batchSize
+    syncCoordinator.pull(repository, lastPullToken, batchSize) { api.pull(batchSize, it).execute().body()!! }
   }
 
   override fun syncConfig(): SyncConfig = config