+++ app/src/main/java/org/simple/clinic/medicalhistory/sync/MedicalHistorySyncApi.kt
@@ -1,7 +1,7 @@
 package org.simple.clinic.medicalhistory.sync
 
-import io.reactivex.Single
 import org.simple.clinic.sync.DataPushResponse
+import retrofit2.Call
 import retrofit2.http.Body
 import retrofit2.http.GET
 import retrofit2.http.Headers
@@ -13,12 +13,12 @@ interface MedicalHistorySyncApi {
   @POST("v3/medical_histories/sync")
   fun push(
       @Body body: MedicalHistoryPushRequest
-  ): Single<DataPushResponse>
+  ): Call<DataPushResponse>
 
   @Headers(value = ["X-RESYNC-TOKEN: 2"])
   @GET("v3/medical_histories/sync")
   fun pull(
       @Query("limit") recordsToPull: Int,
       @Query("process_token") lastPullToken: String? = null
-  ): Single<MedicalHistoryPullResponse>
+  ): Call<MedicalHistoryPullResponse>
 }