+++ app/src/main/java/org/simple/clinic/patient/sync/PatientSyncApi.kt
@@ -1,7 +1,7 @@
 package org.simple.clinic.patient.sync
 
-import io.reactivex.Single
 import org.simple.clinic.sync.DataPushResponse
+import retrofit2.Call
 import retrofit2.http.Body
 import retrofit2.http.GET
 import retrofit2.http.Headers
@@ -13,12 +13,12 @@ interface PatientSyncApi {
   @POST("v3/patients/sync")
   fun push(
       @Body body: PatientPushRequest
-  ): Single<DataPushResponse>
+  ): Call<DataPushResponse>
 
   @Headers(value = ["X-RESYNC-TOKEN: 3"])
   @GET("v3/patients/sync")
   fun pull(
       @Query("limit") recordsToPull: Int,
       @Query("process_token") lastPullTimestamp: String? = null
-  ): Single<PatientPullResponse>
+  ): Call<PatientPullResponse>
 }