+++ app/src/main/java/org/simple/clinic/summary/teleconsultation/sync/TeleconsultFacilityInfoApi.kt
@@ -1,6 +1,6 @@
 package org.simple.clinic.summary.teleconsultation.sync
 
-import io.reactivex.Single
+import retrofit2.Call
 import retrofit2.http.GET
 import retrofit2.http.Headers
 import retrofit2.http.Query
@@ -11,5 +11,5 @@ interface TeleconsultFacilityInfoApi {
   fun pull(
       @Query("limit") recordsToPull: Int,
       @Query("process_token") lastPullToken: String? = null
-  ): Single<TeleconsultFacilityInfoPullResponse>
+  ): Call<TeleconsultFacilityInfoPullResponse>
 }