+++ app/src/main/java/org/simple/clinic/sync/DataSync.kt
@@ -2,14 +2,17 @@ package org.simple.clinic.sync
 
 import io.reactivex.Completable
 import io.reactivex.Observable
+import io.reactivex.Scheduler
+import io.reactivex.Single
+import io.reactivex.rxkotlin.toObservable
+import io.reactivex.schedulers.Schedulers
 import io.reactivex.subjects.PublishSubject
 import org.simple.clinic.di.AppScope
 import org.simple.clinic.platform.analytics.Analytics
 import org.simple.clinic.platform.analytics.SyncAnalyticsEvent
-import org.simple.clinic.platform.analytics.SyncAnalyticsEvent.Completed
-import org.simple.clinic.platform.analytics.SyncAnalyticsEvent.Failed
-import org.simple.clinic.platform.analytics.SyncAnalyticsEvent.Started
 import org.simple.clinic.platform.crash.CrashReporter
+import org.simple.clinic.remoteconfig.RemoteConfigService
+import org.simple.clinic.user.UserSession
 import org.simple.clinic.util.ErrorResolver
 import org.simple.clinic.util.ResolvedError
 import org.simple.clinic.util.ResolvedError.NetworkRelated
@@ -18,15 +21,59 @@ import org.simple.clinic.util.ResolvedError.Unauthenticated
 import org.simple.clinic.util.ResolvedError.Unexpected
 import org.simple.clinic.util.exhaustive
 import org.simple.clinic.util.scheduler.SchedulersProvider
+import org.simple.clinic.util.toOptional
 import timber.log.Timber
+import java.util.concurrent.LinkedBlockingQueue
+import java.util.concurrent.ThreadFactory
+import java.util.concurrent.ThreadPoolExecutor
+import java.util.concurrent.TimeUnit
+import java.util.concurrent.atomic.AtomicInteger
 import javax.inject.Inject
 
+private fun createScheduler(workers: Int): Scheduler {
+  val executor = ThreadPoolExecutor(
+      workers, // number of threads to keep alive always
+      workers, // max number of threads in the pool
+      1L, // time to keep non-core threads alive before shutting down
+      TimeUnit.SECONDS,
+      LinkedBlockingQueue(), // Queue for receiving work
+      SyncThreadFactory()
+  )
+
+  return Schedulers.from(executor)
+}
+
+private class SyncThreadFactory: ThreadFactory {
+
+  private val threadCounter = AtomicInteger(1)
+
+  override fun newThread(job: Runnable): Thread {
+    return Thread(job, "sync-thread-${threadCounter.getAndIncrement()}")
+  }
+}
+
 @AppScope
-class DataSync @Inject constructor(
+class DataSync(
     private val modelSyncs: ArrayList<ModelSync>,
     private val crashReporter: CrashReporter,
-    private val schedulersProvider: SchedulersProvider
+    private val userSession: UserSession,
+    private val schedulersProvider: SchedulersProvider,
+    private val syncScheduler: Scheduler
 ) {
+
+  @Inject constructor(
+      modelSyncs: ArrayList<ModelSync>,
+      crashReporter: CrashReporter,
+      userSession: UserSession,
+      schedulersProvider: SchedulersProvider,
+      remoteConfigService: RemoteConfigService
+  ): this(
+      modelSyncs = modelSyncs,
+      crashReporter = crashReporter,
+      userSession = userSession,
+      schedulersProvider = schedulersProvider,
+      syncScheduler = createScheduler(remoteConfigService.reader().long("max_parallel_syncs", 1L).toInt())
+  )
 
   private val syncProgress = PublishSubject.create<SyncGroupResult>()
 
@@ -41,45 +88,70 @@ class DataSync @Inject constructor(
   }
 
   fun sync(syncGroup: SyncGroup): Completable {
-    return Observable
-        .fromIterable(modelSyncs)
-        .filter { it.syncConfig().syncGroup == syncGroup }
-        .map { modelSync ->
-          val syncName = modelSync.name
+    val syncsInGroup = modelSyncs.filter { it.syncConfig().syncGroup == syncGroup }
 
-          modelSync
-              .sync()
-              .doOnSubscribe { reportSyncEvent(syncName, Started) }
-              .doOnComplete { reportSyncEvent(syncName, Completed) }
-              .doOnError { reportSyncEvent(syncName, Failed) }
+    return Single
+        .fromCallable { userSession.loggedInUserImmediate().toOptional() }
+        .subscribeOn(schedulersProvider.io())
+        .flatMapObservable { user ->
+          syncsInGroup
+              .toObservable()
+              .map { user to it }
         }
+        .filter { (user, modelSync) ->
+          if (modelSync.requiresSyncApprovedUser) {
+            user.isPresent() && user.get().canSyncData
+          } else true
+        }
+        .map { (_, modelSync) -> modelSync }
         .toList()
-        .flatMapCompletable { runAndSwallowErrors(it, syncGroup) }
+        .map(::modelSyncsToCompletables)
+        .flatMapObservable { Observable.fromIterable(it) }
+        .flatMapCompletable { runAndSwallowErrors(it, syncGroup).subscribeOn(syncScheduler) }
         .doOnSubscribe { syncProgress.onNext(SyncGroupResult(syncGroup, SyncProgress.SYNCING)) }
+        .doOnComplete { syncProgress.onNext(SyncGroupResult(syncGroup, SyncProgress.SUCCESS)) }
+        .doOnError { syncProgress.onNext(SyncGroupResult(syncGroup, SyncProgress.FAILURE)) }
   }
 
-  private fun reportSyncEvent(name: String, event: SyncAnalyticsEvent) {
-    Timber.tag("Sync").i("Started sync: $name")
-    Analytics.reportSyncEvent(name, event)
+  private fun modelSyncsToCompletables(modelSyncs: List<ModelSync>): List<Completable> {
+    val allPushes = modelSyncs.map(::generatePushOperationForSync)
+    val allPulls = modelSyncs.map(::generatePullOperationForSync)
+
+    return allPushes + allPulls
+  }
+
+  private fun generatePullOperationForSync(sync: ModelSync): Completable {
+    return Completable
+        .fromAction(sync::pull)
+        .doOnSubscribe { reportSyncEvent(sync.name, "Pull", SyncAnalyticsEvent.Started) }
+        .doOnComplete { reportSyncEvent(sync.name, "Pull", SyncAnalyticsEvent.Completed) }
+        .doOnError { reportSyncEvent(sync.name, "Pull", SyncAnalyticsEvent.Failed) }
+  }
+
+  private fun generatePushOperationForSync(sync: ModelSync): Completable {
+    return Completable
+        .fromAction(sync::push)
+        .doOnSubscribe { reportSyncEvent(sync.name, "Push", SyncAnalyticsEvent.Started) }
+        .doOnComplete { reportSyncEvent(sync.name, "Push", SyncAnalyticsEvent.Completed) }
+        .doOnError { reportSyncEvent(sync.name, "Push", SyncAnalyticsEvent.Failed) }
+  }
+
+  private fun reportSyncEvent(name: String, type: String, event: SyncAnalyticsEvent) {
+    val analyticsName = "$type:$name" // Ex: "Push:Patients"
+    Timber.tag("Sync").i("$analyticsName:${event.name}") // Ex: "Push:Patients:Started"
+    Analytics.reportSyncEvent(analyticsName, event)
   }
 
   fun fireAndForgetSync(syncGroup: SyncGroup) {
-    sync(syncGroup)
-        .subscribeOn(schedulersProvider.io())
-        .subscribe()
+    sync(syncGroup).subscribe()
   }
 
   fun fireAndForgetSync() {
-    syncTheWorld()
-        .subscribeOn(schedulersProvider.io())
-        .subscribe()
+    syncTheWorld().subscribe()
   }
 
-  private fun runAndSwallowErrors(completables: List<Completable>, syncGroup: SyncGroup): Completable {
-    return Completable
-        .mergeDelayError(completables)
-        .doOnComplete { syncProgress.onNext(SyncGroupResult(syncGroup, SyncProgress.SUCCESS)) }
-        .doOnError { syncProgress.onNext(SyncGroupResult(syncGroup, SyncProgress.FAILURE)) }
+  private fun runAndSwallowErrors(completable: Completable, syncGroup: SyncGroup): Completable {
+    return completable
         .doOnError(::logError)
         .onErrorComplete()
   }