+++ app/src/main/java/org/simple/clinic/sync/SyncCoordinator.kt
@@ -1,13 +1,9 @@
 package org.simple.clinic.sync
 
 import com.f2prateek.rx.preferences2.Preference
-import io.reactivex.Completable
-import io.reactivex.Observable
-import io.reactivex.Single
-import io.reactivex.functions.Consumer
 import org.simple.clinic.patient.SyncStatus
-import org.simple.clinic.util.Just
 import org.simple.clinic.util.Optional
+import org.simple.clinic.util.toNullable
 import timber.log.Timber
 import javax.inject.Inject
 
@@ -15,28 +11,30 @@ class SyncCoordinator @Inject constructor() {
 
   fun <T : Any, P> push(
       repository: SynceableRepository<T, P>,
-      pushNetworkCall: (List<T>) -> Single<DataPushResponse>
-  ): Completable = repository
-      .recordsWithSyncStatus(SyncStatus.PENDING)
-      .filter { it.isNotEmpty() }
-      .flatMapSingleElement { pushNetworkCall(it).doOnSuccess(logValidationErrorsIfAny(it)) }
-      .map { it.validationErrors }
-      .map { errors -> errors.map { it.uuid } }
-      .flatMapCompletable { recordIdsWithErrors ->
-        repository
-            .setSyncStatus(from = SyncStatus.PENDING, to = SyncStatus.DONE)
-            .andThen(when {
-              recordIdsWithErrors.isEmpty() -> Completable.complete()
-              else -> repository.setSyncStatus(recordIdsWithErrors, SyncStatus.INVALID)
-            })
-      }
+      pushNetworkCall: (List<T>) -> DataPushResponse
+  ) {
+    val pendingSyncRecords = repository.recordsWithSyncStatus(SyncStatus.PENDING)
 
-  private fun <T : Any> logValidationErrorsIfAny(records: List<T>): Consumer<in DataPushResponse> {
-    return Consumer { response ->
-      if (response.validationErrors.isNotEmpty()) {
-        val recordType = records.first().javaClass.simpleName
-        Timber.e("Server sent validation errors when syncing $recordType : ${response.validationErrors}")
-      }
+    if (pendingSyncRecords.isNotEmpty()) {
+      val response = pushNetworkCall(pendingSyncRecords)
+      repository.setSyncStatus(SyncStatus.PENDING, SyncStatus.DONE)
+
+      val validationErrors = response.validationErrors
+      handleValidationErrors(validationErrors, pendingSyncRecords, repository)
+    }
+  }
+
+  private fun <P, T : Any> handleValidationErrors(
+      validationErrors: List<ValidationErrors>,
+      pendingSyncRecords: List<T>,
+      repository: SynceableRepository<T, P>
+  ) {
+    val recordIdsWithErrors = validationErrors.map { it.uuid }
+    if (recordIdsWithErrors.isNotEmpty()) {
+      val recordType = pendingSyncRecords.first().javaClass.simpleName
+      Timber.e("Server sent validation errors when syncing $recordType : $validationErrors")
+
+      repository.setSyncStatus(recordIdsWithErrors, SyncStatus.INVALID)
     }
   }
 
@@ -44,18 +42,19 @@ class SyncCoordinator @Inject constructor() {
       repository: SynceableRepository<T, P>,
       lastPullToken: Preference<Optional<String>>,
       batchSize: Int,
-      pullNetworkCall: (String?) -> Single<out DataPullResponse<P>>
-  ): Completable {
-    return lastPullToken.asObservable()
-        .take(1)
-        .flatMapSingle { (lastPull) -> pullNetworkCall(lastPull) }
-        .flatMap { response ->
-          repository.mergeWithLocalData(response.payloads)
-              .andThen(Completable.fromAction { lastPullToken.set(Just(response.processToken)) })
-              .andThen(Observable.just(response))
-        }
-        .repeat()
-        .takeWhile { response -> response.payloads.size >= batchSize }
-        .ignoreElements()
+      pullNetworkCall: (String?) -> DataPullResponse<P>
+  ) {
+    var hasFetchedAllData = false
+
+    while (!hasFetchedAllData) {
+      val processToken = lastPullToken.get().toNullable()
+
+      val response = pullNetworkCall(processToken)
+
+      repository.mergeWithLocalData(response.payloads)
+      lastPullToken.set(Optional.of(response.processToken))
+
+      hasFetchedAllData = response.payloads.size < batchSize
+    }
   }
 }