+++ app/src/main/java/org/simple/clinic/sync/SyncSetup.kt
@@ -4,7 +4,6 @@ import io.reactivex.disposables.CompositeDisposable
 import io.reactivex.disposables.Disposable
 import org.simple.clinic.protocol.SyncProtocolsOnLogin
 import org.simple.clinic.sync.indicator.SyncIndicatorStatusCalculator
-import org.simple.clinic.util.scheduler.SchedulersProvider
 import javax.annotation.CheckReturnValue
 import javax.inject.Inject
 
@@ -13,8 +12,7 @@ class SyncSetup @Inject constructor(
     private val syncScheduler: SyncScheduler,
     private val syncIndicatorStatusCalculator: SyncIndicatorStatusCalculator,
     private val syncProtocolsOnLogin: SyncProtocolsOnLogin,
-    private val dataSyncOnApproval: IDataSyncOnApproval,
-    private val schedulersProvider: SchedulersProvider
+    private val dataSyncOnApproval: IDataSyncOnApproval
 ) {
 
   @CheckReturnValue
@@ -25,7 +23,7 @@ class SyncSetup @Inject constructor(
           dataSyncOnApproval.sync(),
           syncScheduler.schedule().subscribe(),
           syncIndicatorStatusCalculator.updateSyncResults(),
-          dataSync.syncTheWorld().subscribeOn(schedulersProvider.io()).subscribe()
+          dataSync.syncTheWorld().subscribe()
       )
     }
   }